fix publishing
