+++ web/src/beta/lib/core/StoryPanel/Page/index.tsx
@@ -87,7 +87,7 @@ const StoryPanel: React.FC<Props> = ({
       onClickAway={onPageSelect}
       onSettingsToggle={onPageSettingsToggle}>
       <Wrapper id={page?.id} padding={panelSettings.padding.value} gap={panelSettings.gap.value}>
-        {(isEditable || title?.title?.value) && (
+        {(isEditable ?? title?.title?.value) && (
           <StoryBlock
             block={{
               id: titleId,
